backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
cryptography==41.0.5
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aioimaplib==1.0.1
email-validator==2.1.0.post1
bcrypt==4.0.1
requests==2.31.0
python-jose[cryptography]==3.3.0 